Implement MultiFilePicker for MultiFileDataField.
A two page file picker that presents the user with a list of currently
active files on the first page and a set of buttons to add or remove
files from the active list. Adding a file calls the normal FilePicker()
on the second page to pick a single file from a list of all files and
adds it to the active selection.

Internally uses a ListPickerWidget, so the class definition of the
ListPickerWidget class was moved to ListPicker.hpp