Implement MultiFilePicker for MultiFileDataField
A two page file picker that presents the user with a list of currently
active files on the first page together with a set of buttons to add and
remove files from the list.
Pressing "Add" opens a FilePicker on the FileDataField contained inside
the associated MultiFileDataField, which presents the user with a list of
all the airspace files found in the directory (both active and unactive).
The selected file then gets added to the selection inside the associated
MultiFileDataField.
Pressing "Remove" causes the selected file to be removed from the active
selection in the associated MultiFileDataField.